Move ConcatLibrary to feature_library.py and add a __add__ method to baseclass.
